Move interactiveSubprocess to Utils
