+++ main/src/mill/modules/Util.scala
@@ -1,8 +1,9 @@
 package mill.modules
 
-
-import ammonite.ops.{Path, RelPath, empty, mkdir, read}
+import java.io.{ByteArrayInputStream}
+import ammonite.ops.{Path, RelPath, InteractiveShelloutException, empty, mkdir, read}
 import mill.eval.PathRef
+import mill.clientserver.InputPumper
 import mill.util.{Ctx, IO}
 
 object Util {
@@ -53,5 +54,44 @@ object Util {
       }
     })()
     PathRef(ctx.dest / dest)
+  }
+
+  def interactiveSubprocess(commandArgs: Seq[String],
+                            envArgs: Map[String, String],
+                            workingDir: Path) = {
+    val builder = new java.lang.ProcessBuilder()
+
+    for ((k, v) <- envArgs){
+      if (v != null) builder.environment().put(k, v)
+      else builder.environment().remove(k)
+    }
+    builder.directory(workingDir.toIO)
+
+    val process = if (System.in.isInstanceOf[ByteArrayInputStream]){
+
+      val process = builder
+        .command(commandArgs:_*)
+        .start()
+
+      val sources = Seq(
+        process.getInputStream -> System.out,
+        process.getErrorStream -> System.err,
+        System.in -> process.getOutputStream
+      )
+
+      for((std, dest) <- sources){
+        new Thread(new InputPumper(std, dest, false)).start()
+      }
+      process
+    } else {
+      builder
+        .command(commandArgs:_*)
+        .inheritIO()
+        .start()
+    }
+
+    val exitCode = process.waitFor()
+    if (exitCode == 0) ()
+    else throw InteractiveShelloutException()
   }
 }
